requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
ccxt==4.4.17
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
numpy==2.1.2
pandas==2.2.3
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
yarl==1.15.1
